refactor: move external utils to StakeWise file
